+++ src/components/Directions.tsx
@@ -1,0 +1,14 @@
+import React, {useRef, useState, useEffect} from 'react';
+
+
+function Directions() {
+        
+    return (
+        <div className='directions-section'>
+            
+            
+        </div>
+    );
+}
+
+export default Directions;
